chore(structure): restructured project layout to be more modular (#81)
