+++ src/app/app.component.ts
@@ -2,8 +2,7 @@ import { Component } from '@angular/core';
 
 @Component({
     selector: 'app-root',
-    templateUrl: './app.component.html',
-    styleUrls: [ './app.component.scss' ],
+    template: '<router-outlet></router-outlet>',
 })
 export class AppComponent {
 }
