+++ src/app/routing.module.ts
@@ -1,28 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { ErrorComponent } from './error/error.component';
-import { LandingComponent } from './landing/landing.component';
-import { ScheduleComponent } from './schedule/schedule.component';
-
-import { ScheduleServiceResolve } from '../services/schedule.service';
-
 const routes: Routes = [
     {
         path: '',
-        pathMatch: 'full',
-        component: LandingComponent
-    },
-    {
-        path: 'schedule',
-        component: ScheduleComponent,
-        resolve: {
-            scheduleData: ScheduleServiceResolve
-        }
-    },
-    {
-        path: '**',
-        component: ErrorComponent,
+        loadChildren: './site/site.module#SiteModule',
     },
 ];
