+++ src/app/site/site.component.html
@@ -14,7 +14,7 @@
     <div class="navigation" role="navigation">
         <div class="container">
             <a [routerLink]="['/']" routerLinkActive="active" [routerLinkActiveOptions]="{ exact: true }">Home</a>
-            <a [routerLink]="['schedule']" routerLinkActive="active">Schedule</a>
+            <a [routerLink]="['/schedule']" routerLinkActive="active">Schedule</a>
         </div>
     </div>
