+++ src/app/site/site.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+    templateUrl: './site.component.html',
+    styleUrls: [ './site.component.scss' ],
+})
+export class SiteComponent {
+}
